# ui/public/openapi.yaml
openapi: 3.1.0
info:
  title: LV Trade Platform API
  version: 0.1.0
servers:
  - url: https://api.example.com
paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: OK
  /v1/auth/register:
    post:
      summary: Register and receive token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /v1/auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/me:
    get:
      security:
        - bearerAuth: []
      summary: Current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/balances:
    get:
      security:
        - bearerAuth: []
      summary: Balances
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Balance"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/faucet:
    post:
      security:
        - bearerAuth: []
      summary: Faucet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaucetRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/orders:
    post:
      security:
        - bearerAuth: []
      summary: Place order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceOrderRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaceOrderResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/market/candles:
    get:
      summary: Market candles
      parameters:
        - name: pair
          in: query
          required: true
          schema:
            type: string
        - name: timeframe
          in: query
          required: true
          schema:
            type: string
            enum: [1m, 5m, 10m, 15m, 30m, 1h]
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: fresh
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candle"
        "400":
          $ref: "#/components/responses/BadRequest"
  /v1/market/candles/ws:
    get:
      summary: Market candles WebSocket
      parameters:
        - name: pair
          in: query
          required: true
          schema:
            type: string
        - name: timeframe
          in: query
          required: true
          schema:
            type: string
            enum: [1m, 5m, 10m, 15m, 30m, 1h]
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: fresh
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "101":
          description: Switching Protocols
  /v1/internal/deposits:
    post:
      summary: Internal deposit
      parameters:
        - name: X-Internal-Token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InternalMovement"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/internal/withdrawals:
    post:
      summary: Internal withdrawal
      parameters:
        - name: X-Internal-Token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InternalMovement"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user_id:
          type: string
        access_token:
          type: string
      required: [user_id, access_token]
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
      required: [access_token]
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
      required: [id, email]
    Balance:
      type: object
      properties:
        asset_id:
          type: string
        symbol:
          type: string
        kind:
          type: string
        amount:
          type: string
      required: [asset_id, symbol, kind, amount]
    PlaceOrderRequest:
      type: object
      properties:
        pair:
          type: string
        side:
          type: string
          enum: [buy, sell]
        type:
          type: string
          enum: [limit, market]
        price:
          type: string
        qty:
          type: string
        quote_amount:
          type: string
        time_in_force:
          type: string
          enum: [gtc, ioc]
        client_ref:
          type: string
      required: [pair, side, type, time_in_force]
    FaucetRequest:
      type: object
      properties:
        asset:
          type: string
        amount:
          type: string
        reference:
          type: string
      required: [asset, amount]
    PlaceOrderResponse:
      type: object
      properties:
        order_id:
          type: string
        status:
          type: string
      required: [order_id, status]
    InternalMovement:
      type: object
      properties:
        user_id:
          type: string
        asset:
          type: string
        amount:
          type: string
        reference:
          type: string
      required: [user_id, asset, amount, reference]
    StatusResponse:
      type: object
      properties:
        status:
          type: string
      required: [status]
    Candle:
      type: object
      properties:
        time:
          type: integer
        open:
          type: string
        high:
          type: string
        low:
          type: string
        close:
          type: string
      required: [time, open, high, low, close]